requirements.txt
Django==5.2.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.0
django-filter==25.1
django-cors-headers==4.7.0
dj-database-url==3.0.0
PyMySQL==1.1.0
python-dotenv==1.1.1
whitenoise==6.7.0
gunicorn==23.0.0
psycopg2-binary

# keep these three to satisfy any legacy DRF schema imports, just in case
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.1.1

# (rest of your PDF/report libs can stay if you use them)
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cairocffi==1.7.1
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
et_xmlfile==2.0.0
fonttools==4.58.4
html5lib==1.1
idna==3.10
lxml==5.4.0
numpy==1.26.4
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.0
Pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.4
setuptools==74.1.2
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1